refactor: change folders structure
