+++ src/components/playground/requestEditor/RequestEditor.tsx
@@ -1,6 +1,6 @@
 import CodeMirror from '@uiw/react-codemirror';
 import { json } from '@codemirror/lang-json';
-import { useAppSelector } from '@/app/lib/redux/hooks/redux';
+import { useAppSelector } from '@/lib/redux/hooks/redux';
 
 export default function RequestEditor() {
   const requestJson = useAppSelector((state) => state.requestJson.requestJson);
